+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Release Notes
 =============
 
-## Next Releases
+## 0.29.0
 
 **New**